+++ CHANGELOG.md
@@ -57,6 +57,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Allow installations up to PHP 8.0 (#758)
 
 ### Changed
+- !!! Namespace all classes (#1004)
 - Document the required MySQL/MariaDB settings for the FE editor (#843)
 - Simplify the user and group mapper inheritance chains (#806)
 - Move more tests to the nimut testing framework (#804)